big edits! sorry i didn't commit earlier

+++ lib/models/highscores_data.dart
@@ -6,7 +6,7 @@ class HighscoresData {
   final DateTime updatedAt;
   final int deckId;
   final int userId;
-  final double time;
+  final int time;
   final User user;
 
   HighscoresData({
